Update New Politics

+++ Util/Politic/VoteResult.cs
@@ -8,22 +8,22 @@
 		/// <summary>
 		/// 同意
 		/// </summary>
-		public byte Agree { get; }
+		public int Agree { get; }
 		/// <summary>
 		/// 反对
 		/// </summary>
-		public byte Disagree { get; }
+		public int Disagree { get; }
 		/// <summary>
 		/// 弃权
 		/// </summary>
-		public byte NoVote { get; }
+		public int NoVote { get; }
 		/// <summary>
 		/// 投票结果
 		/// </summary>
 		/// <param name="agree">同意</param>
 		/// <param name="disagree">反对</param>
 		/// <param name="noVote">弃权</param>
-		public VoteResult(byte agree, byte disagree, byte noVote) {
+		public VoteResult(int agree, int disagree, int noVote) {
 			this.Agree = agree;
 			this.Disagree = disagree;
 			this.NoVote = noVote;
